+++ src/frontend/apps/impress/src/features/docs/doc-share/api/useCreateDocInvitation.tsx
@@ -4,7 +4,6 @@ import { APIError, errorCauses, fetchAPI } from '@/api';
 import { User } from '@/features/auth';
 import { Doc, Role } from '@/features/docs/doc-management';
 import { Invitation, OptionType } from '@/features/docs/doc-share/types';
-import { ContentLanguage } from '@/i18n/types';
 
 import { KEY_LIST_DOC_INVITATIONS } from './useDocInvitations';
 
@@ -12,20 +11,15 @@ interface CreateDocInvitationParams {
   email: User['email'];
   role: Role;
   docId: Doc['id'];
-  contentLanguage: ContentLanguage;
 }
 
 export const createDocInvitation = async ({
   email,
   role,
   docId,
-  contentLanguage,
 }: CreateDocInvitationParams): Promise<Invitation> => {
   const response = await fetchAPI(`documents/${docId}/invitations/`, {
     method: 'POST',
-    headers: {
-      'Content-Language': contentLanguage,
-    },
     body: JSON.stringify({
       email,
       role,